+++ noodles-bcf/src/reader/header.rs
@@ -1,7 +1,4 @@
-use std::{
-    ffi::CStr,
-    io::{self, Read},
-};
+use std::io::{self, BufRead, BufReader, Read};
 
 use byteorder::{LittleEndian, ReadBytesExt};
 use noodles_vcf as vcf;
@@ -12,38 +9,60 @@ pub(super) fn read_header<R>(reader: &mut R) -> io::Result<(vcf::Header, StringM
 where
     R: Read,
 {
-    let raw_header = read_raw_header(reader)?;
+    const NUL: u8 = 0x00;
 
-    let header = raw_header
-        .parse()
-        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
+    let l_text = reader.read_u32::<LittleEndian>().map(u64::from)?;
 
-    let string_maps = raw_header
-        .parse()
+    let mut parser = vcf::header::Parser::default();
+    let mut string_maps = StringMaps::default();
+
+    let mut header_reader = BufReader::new(reader.take(l_text));
+    let mut buf = Vec::new();
+
+    while read_line(&mut header_reader, &mut buf)? != 0 {
+        if buf == [NUL] {
+            break;
+        }
+
+        let entry = parser
+            .parse_partial(&buf)
+            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
+
+        string_maps
+            .insert_entry(&entry)
+            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
+    }
+
+    let header = parser
+        .finish()
         .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
 
     Ok((header, string_maps))
 }
 
-pub fn read_raw_header<R>(reader: &mut R) -> io::Result<String>
+fn read_line<R>(reader: &mut R, dst: &mut Vec<u8>) -> io::Result<usize>
 where
-    R: Read,
+    R: BufRead,
 {
-    let l_text = reader.read_u32::<LittleEndian>().and_then(|n| {
-        usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-    })?;
+    const LINE_FEED: u8 = b'\n';
+    const CARRIAGE_RETURN: u8 = b'\r';
 
-    let mut buf = vec![0; l_text];
-    reader.read_exact(&mut buf)?;
+    dst.clear();
 
-    CStr::from_bytes_with_nul(&buf)
-        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-        .and_then(|c_header| {
-            c_header
-                .to_str()
-                .map(|s| s.into())
-                .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-        })
+    match reader.read_until(LINE_FEED, dst)? {
+        0 => Ok(0),
+        n => {
+            if dst.ends_with(&[LINE_FEED]) {
+                dst.pop();
+
+                if dst.ends_with(&[CARRIAGE_RETURN]) {
+                    dst.pop();
+                }
+            }
+
+            Ok(n)
+        }
+    }
 }
 
 #[cfg(test)]
@@ -51,19 +70,30 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_read_raw_header() -> io::Result<()> {
+    fn test_read_header() -> io::Result<()> {
+        use vcf::header::FileFormat;
+
         const NUL: u8 = 0x00;
 
-        let raw_header = "##fileformat=VCFv4.3\n";
+        let raw_header = b"##fileformat=VCFv4.3
+#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO
+";
 
-        let mut data = 22u32.to_le_bytes().to_vec(); // l_text = 22
-        data.extend_from_slice(raw_header.as_bytes());
+        let mut data = 60u32.to_le_bytes().to_vec(); // l_text = 22
+        data.extend_from_slice(raw_header);
         data.push(NUL);
 
         let mut reader = &data[..];
-        let actual = read_raw_header(&mut reader)?;
+        let (actual_header, actual_string_maps) = read_header(&mut reader)?;
 
-        assert_eq!(actual, raw_header);
+        let expected_header = vcf::Header::builder()
+            .set_file_format(FileFormat::new(4, 3))
+            .build();
+
+        let expected_string_maps = StringMaps::default();
+
+        assert_eq!(actual_header, expected_header);
+        assert_eq!(actual_string_maps, expected_string_maps);
 
         Ok(())
     }